requirements.txt
streamlit
PyMuPDF==1.23.6
Pillow
pandas
cohere
numpy
